requirements.txt
opencv_contrib_python==4.1.2.30
numpy==1.18.0
torch==1.3.1
torchvision==0.4.2
scikit_image==0.16.2
Pillow==7.0.0
skimage==0.0
